pyproject.toml
[tool.poetry]
name = "fastapi-asgi"
version = "0.1.0"
description = "api setup to run asgi type app"
authors = ["gsubramanian <gsubramanian@tesla.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
gunicorn = "^21.2.0"
bson = "^0.5.10"
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
greenlet = "^3.0.3"
requests = "^2.31.0"
pyjwt = "^2.8.0"
cryptography = "^42.0.5"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.5"
black = "^24.3.0"
mypy = "^1.9.0"
pytest = "^8.1.1"
coverage = "^7.4.4"
pytest-httpx = "^0.30.0"
pytest-asyncio = "^0.23.6"

[tool.black]
line-length = 88

[tool.ruff]
select = ["E", "F", "I"]
fixable = ["ALL"]
exclude = [".git", ".mypy_cache", ".ruff_cache"]
line-length = 88

[tool.mypy]
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
disallow_untyped_decorators = false
ignore_missing_imports = true
implicit_reexport = true
plugins = "pydantic.mypy"

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
